Import the asset packs for GarbagePoints and recycle dialog
Add the assets from unity store that are going to be used for designing the garbage points as points of interest and buttons of the recycle dialog. 3D exclamation mark will point to the GarbagePoints to make them look like points of interest. Recycle dialog will have buttons for navigation.

Assets packs:
- Prototyping Pack - Prefab for exclamation mark
- TextMesh Pro - Buttons and text